% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars2.R
\docType{data}
\name{cars2}
\alias{cars2}
\title{Speed and Stopping Distances of Cars}
\format{
A data frame with 50 observations on 2 variables
/itemize{
/item [,1] speed numeric (mph)
/item [,2] distance numeric (ft)
}
}
\source{
Ezekiel, M. (1930) Methods of Correlation Analysis. Wiley.
}
\usage{
cars2
}
\description{
The data give the speed of cars and the distances taken to stop.
}
\keyword{datasets}
